added similarity check to keyword search, moved get_keyword_from_synonyms and get_keyword to helper_service.py
