+++ src/nlp/src/api/helper_service.py
@@ -1,5 +1,18 @@
-from array import array
+import logging
+import sys
+
+logging.basicConfig(format='%(asctime)s %(levelname)s %(name)s : %(message)s',
+                    datefmt='%d.%m.%Y %H:%M:%S',
+                    encoding='utf-8',
+                    stream=sys.stdout,
+                    level=logging.INFO)
+LOGGER = logging.getLogger("[STRING INTERPRETER][HELPER SERVICE]")
+
 import numpy as np
+from array import array
+
+from api.utils import get_entity_synonyms, get_alias_synonyms
+
 
 def convert_number_to_meter(unit: str, number: int) -> int:
     """
@@ -71,4 +84,43 @@ def compute_levenshtein_distance(token_1: str, token_2: str)-> int:
     # compute final Levenshtein Distance Ratio
     distance_ratio = ((len(token_1)+len(token_2)) - distance[row][col]) / (len(token_1)+len(token_2))
     return distance_ratio
- 
+
+
+def get_keyword_from_synonyms(string: str, default_keyword: str, synonyms: dict):
+    """
+    checks if a word matches to a synonym from a given dictionary
+    :param  string token to be allocated to a synonym class
+    :param  default_keyword the keyword used in case if no keyword was found
+    :param  synonyms dictionary with keywords and their synonyms/aliases
+    :return key synonym if the word is in the synonym list, otherwise default_keyword for the synonym class
+    """
+    # get keyword
+    for keyword in synonyms.keys():
+        # iterate over all items in the synonym list and check if token is equal to one synonym
+        for synonym in synonyms[keyword]:
+            if check_similarity(synonym, string.lower(), threshold=0.85):
+                LOGGER.debug("Found matching keyword %s for %s", keyword, string)
+                return keyword
+
+    LOGGER.info(
+        "Couldn't find a matching keyword for %s, "
+        "using default keyword %s",
+        string, default_keyword
+    )
+    return default_keyword
+
+
+def get_keyword(string: str, default_keyword: str, synonym_class: str, is_entity: bool) -> str:
+    """
+    checks if a word matches to a synonym from a specific class
+    :param  string token to be allocated to a synonym class
+    :param  default_keyword the keyword used in case if no keyword was found
+    :param  synonym_class the class in which the synonyms are searched for
+    :param  is_entity defines whether synonym_class is a named entity
+    :return key synonym if the word is in the synonym list, otherwise default_keyword for the synonym class
+    """
+    if is_entity:
+        synonyms = get_entity_synonyms(entity=synonym_class)
+    else:
+        synonyms = get_alias_synonyms(title=synonym_class)
+    return get_keyword_from_synonyms(string, default_keyword, synonyms[synonym_class])
