+++ src/nlp/src/api/utils.py
@@ -7,6 +7,7 @@ logging.basicConfig(format='%(asctime)s %(levelname)s %(name)s : %(message)s',
                     stream=sys.stdout,
                     level=logging.INFO)
 LOGGER = logging.getLogger("[STRING INTERPRETER][UTILS]")
+
 import itertools
 import os
 import pathlib
@@ -157,40 +158,3 @@ def get_alias_synonyms(chatette_file_path: str = None,
     except TypeError:
         return aliases
     return aliases
-
-
-def get_keyword_from_synonyms(string: str, default_keyword: str, synonyms: dict):
-    """
-    checks if a word matches to a synonym from a given dictionary
-    :param  string token to be allocated to a synonym class
-    :param  default_keyword the keyword used in case if no keyword was found
-    :param  synonyms dictionary with keywords and their synonyms/aliases
-    :return key synonym if the word is in the synonym list, otherwise default_keyword for the synonym class
-    """
-    # get keyword
-    for keyword in synonyms.keys():
-        if string.lower() in synonyms[keyword]:
-            LOGGER.debug(f"Found matching keyword {keyword} for {string}")
-            return keyword
-
-    LOGGER.info(
-        f"Couldn't find a matching keyword for {string}, "
-        f"using default keyword {default_keyword}"
-    )
-    return default_keyword
-
-
-def get_keyword(string: str, default_keyword: str, synonym_class: str, is_entity: bool) -> str:
-    """
-    checks if a word matches to a synonym from a specific class
-    :param  string token to be allocated to a synonym class
-    :param  default_keyword the keyword used in case if no keyword was found
-    :param  synonym_class the class in which the synonyms are searched for
-    :param  is_entity defines whether synonym_class is a named entity
-    :return key synonym if the word is in the synonym list, otherwise default_keyword for the synonym class
-    """
-    if is_entity:
-        synonyms = get_entity_synonyms(entity=synonym_class)
-    else:
-        synonyms = get_alias_synonyms(title=synonym_class)
-    return get_keyword_from_synonyms(string, default_keyword, synonyms[synonym_class])
